+++ src/LogDimensionCollection.AspNetCore/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [1.2.0] - 2021-05-20
 
+### Added
+
+- `ActionArgDimensionSelector`: new option to auto collect
+
 ### Fixes
 
 - `DimensionsCollectionHelper` should serialize Newtonsoft specific types